requirements.txt
aiohttp
PyYAML
pysolarmanv5==3.0.6